requirements.txt
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
flask==2.3.2
flask-cors==4.0.0
pillow==10.0.0
opencv-python-headless==4.8.0.74
locust==2.15.1
gunicorn==21.2.0
werkzeug==2.3.6